central/requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Pypeul==0.3.1
bottle==0.12.7
feedparser==5.1.3
pycrypto==2.6.1
requests==2.3.0
